Create AbsoluteDifference.py
You are given an array of integers and you need to find out if the absolute difference of values of any two consecutive array integers is at-most D. Input: First line of input contains an integer T, denoting the number of test cases. You have to give your output for every test case. Every test case contains the size of the array(N), the absolute value difference that you have to check for (D) and the values of the array itself. First line of each test case contains two integers N and D, denoting the size of the array and value of D respectively. Second line of each test case contains N integers, which are separated by one space and are the integers for the array.

output :
print YES or NO